Rename modules directory to hitl
