+++ hardware_in_the_loop/software/docs/source/index.rst
@@ -3,7 +3,7 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to hardware_in_the_loop's documentation!
+Welcome to the OlinHITL documentation!
 ================================================
 
 .. toctree::
